Se implementaron las funciones para imprimir los tiempos y se probaron las intancias de prueba. SE comprobo que funciona correctamente la funcion set_times, sin embargo falta agregar la consolidacion en los CD lo cual es avance de memoria, no de IAA.

+++ GRASP VRPCD/vehicle.h
@@ -16,7 +16,7 @@ class Vehicle{
 		// Lista de la cantidad items que contiene hasta el momento de cada supplier de la forma [{10,1},{11,2}...] 10 elem, del suplier 1, 11 del 2, etc.
 		vector<tuple<int,int>> items;
 		// Lista de tiempos de llegada y salida en cada nodo de la forma [{t_llegada1,t_salida1},{t_llegada2,t_salida2}...]
-		vector<tuple<int,int>> pickup_times,crossdock_times,delivery_times;
+		vector<tuple<float,float>> pickup_times,crossdock_times,delivery_times;
 
 		// funciones de la clase
 		Vehicle(int,int,int);
